Move share button into share component
